Rename flo coordinator module
