+++ homeassistant/components/flo/switch.py
@@ -14,7 +14,7 @@ from homeassistant.helpers import entity_platform
 from homeassistant.helpers.entity_platform import AddEntitiesCallback
 
 from .const import DOMAIN as FLO_DOMAIN
-from .device import FloDeviceDataUpdateCoordinator
+from .coordinator import FloDeviceDataUpdateCoordinator
 from .entity import FloEntity
 
 ATTR_REVERT_TO_MODE = "revert_to_mode"
